+++ janusgraph-core/src/main/java/org/janusgraph/core/JanusGraph.java
@@ -52,6 +52,11 @@ import org.janusgraph.graphdb.database.cache.CacheInvalidationService;
         reason = "TinkerPop assumes cardinality is SINGLE when not explicitly given, while JanusGraph uses the cardinality " +
             "already defined in the schema.")
 @Graph.OptOut(
+    test = "org.apache.tinkerpop.gremlin.structure.TransactionMultiThreadedTest",
+    method = "*",
+    reason = "TinkerPop assumes certain behaviour of concurrent transactions that JanusGraph does not follow, including " +
+        "1) Check vertex and edge existence before committing; 2) Conflicting transactions should abort and throw TransactionException.")
+@Graph.OptOut(
         test = "org.apache.tinkerpop.gremlin.process.computer.GraphComputerTest",
         method = "shouldOnlyAllowReadingVertexPropertiesInMapReduce",
         reason = "JanusGraph simply throws the wrong exception -- should not be a ReadOnly transaction exception but a specific one for MapReduce. This is too cumbersome to refactor in JanusGraph.")
@@ -86,11 +91,6 @@ import org.janusgraph.graphdb.database.cache.CacheInvalidationService;
         test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.MergeVertexTest",
         method = "g_withSideEffectXc_label_person_name_markoX_withSideEffectXm_age_19X_mergeVXselectXcXX_optionXonMatch_selectXmXX_option",
         reason = "JanusGraph doesn't support MergeVertex yet.")
-@Graph.OptOut(
-        test = "org.apache.tinkerpop.gremlin.process.traversal.step.map.CountTest",
-        method = "g_VX1X_valuesXageX_countXlocalX",
-        reason = "TinkerPop 3.6.4 buggy test. We should remove this OptOut when updating to TinkerPop >= 3.6.5 because the test was fixed in TinkerPop 3.6.5. " +
-            "See comment here: https://github.com/JanusGraph/janusgraph/pull/3780#issuecomment-1571961290")
 public interface JanusGraph extends Transaction {
 
     /**